+++ src/components/ButtonLabel.jsx
@@ -1,24 +1,31 @@
 import React from 'react'
 import styled from "styled-components";
 
-const ButtonWrapper = styled.section`
+import PropTypes from 'prop-types';
+
+
+
+
+export const ButtonLabel = ({buttonText,buttonWidth,buttonHeight,buttonColor,...args}) => {
+
+    const ButtonWrapper = styled.section`
 
 align-items: center;
 padding: 12px 24px;
 display: flex;
 flex-direction: row;
+text-align:center;
 
-padding: 12px 24px;
 
 position: static;
-width: 135px;
-height: 40px;
+width: ${buttonWidth};
+height: ${buttonHeight};
 left: 0px;
 top: 200px;
 
 /* Primary/600 */
-
-background: #0E5E81;
+color: white;
+background: ${buttonColor} ;
 /* Button 1 */
 
 box-shadow: 0px 4px 8px rgba(34, 44, 47, 0.16);
@@ -38,12 +45,23 @@ margin: 16px 0px;
 
 `;
 
-const ButtonLabel = () => {
+
     return (
        
-        <button> <ButtonWrapper>Sign me up!</ButtonWrapper></button>
+        <button> <ButtonWrapper>{buttonText}</ButtonWrapper></button>
 
     )
 }
+
+ButtonLabel.propTypes = {
+    buttonColor: PropTypes.string.isRequired,
+    buttonText: PropTypes.string.isRequired,
+    buttonWidth: PropTypes.string.isRequired,
+    buttonWheight: PropTypes.string.isRequired,
+    
+    
+    
+};
+
 
 export default ButtonLabel;
