+++ src/stories/ReviewCard.stories.js
@@ -1,26 +1,26 @@
 import React from 'react';
-import {ReviewCardForStorybook} from './ReviewCardForStorybook';
-import {SliderForStorybook} from './SliderForStorybook';
+import ReviewCard from '../components/ReviewCard';
+import Slider from '../components/Slider';
 //import { storiesOf } from '@storybook/react';
-//import ApolloWraper from '../src/components/ReviewCardForStorybook/ApolloProvider';
+//import ApolloWraper from '../src/components/ReviewCard/ApolloProvider';
 //import { addDecorator } from "@storybook/react";
 
 export default {
      title: "Components",
-     component: SliderForStorybook,
-     subcomponents: { ReviewCardForStorybook },
+     component: Slider,
+     subcomponents: { ReviewCard },
     };
 const Template = (args) => (
-    <SliderForStorybook {...args}>
+    <Slider {...args}>
       
-    </SliderForStorybook>
+    </Slider>
 );
   export const reviewCard = Template.bind({});
   reviewCard.args = {
     description: 'This course provides a great overview of Blender, an extermely powerful (and also incredibly complicated) 3D modeling and animation program. Mikey, the main instructor, is amusing and informative. Both he and Ben student work',
     lname: 'Maurica',
     fname: 'Alejandro',
-    title: 'Ecosystem Builder, Barcelon',
+    title: 'Ecosystem Builder, Barcelona',
     intervalMS: 6100,
     showArrows:true,
   };
